bot/requirement.txt
python-dotenv==1.0.0
asyncio==3.4.3
aiogram==3.3.0
aioredis==2.0.1
aiohttp==3.8.4
redis==4.5.4
psycopg2==2.9.6
